Extracting WordPress dependencies
- Moving most @WordPress external dependencies to the calypso-build package.
- babel-related dependencies stay in the main package.json file as they seem too interwoven and splitting them away requires more work than expected, so it requires a separate issue.
