+++ webpack.config.js
@@ -18,6 +18,7 @@ const MomentTimezoneDataPlugin = require( 'moment-timezone-data-webpack-plugin' 
 const Minify = require( '@automattic/calypso-build/webpack/minify' );
 const SassConfig = require( '@automattic/calypso-build/webpack/sass' );
 const TranspileConfig = require( '@automattic/calypso-build/webpack/transpile' );
+const wordpressExternals = require( '@automattic/calypso-build/webpack/wordpress-externals' );
 
 /**
  * Internal dependencies
@@ -89,32 +90,6 @@ function createProgressHandler() {
 		console.log( message ); // eslint-disable-line no-console
 	};
 }
-
-/**
- * Converts @wordpress require into window reference
- *
- * Note this isn't the same as camel case because of the
- * way that numbers don't trigger the capitalized next letter
- *
- * @example
- * wordpressRequire( '@wordpress/api-fetch' ) = 'wp.apiFetch'
- * wordpressRequire( '@wordpress/i18n' ) = 'wp.i18n'
- *
- * @param {string} request import name
- * @return {string} global variable reference for import
- */
-const wordpressRequire = request => {
-	// @wordpress/components -> [ @wordpress, components ]
-	const [ , name ] = request.split( '/' );
-
-	// components -> wp.components
-	return `wp.${ name.replace( /-([a-z])/g, ( match, letter ) => letter.toUpperCase() ) }`;
-};
-
-const wordpressExternals = ( context, request, callback ) =>
-	/^@wordpress\//.test( request )
-		? callback( null, `root ${ wordpressRequire( request ) }` )
-		: callback();
 
 /**
  * Return a webpack config object
